requirements.txt
Pillow==10.2.0
streamlit==1.29.0
ultralytics==8.1.5
